requirements.txt
Flask==0.10.1
Jinja2==2.7.3
PyYAML==3.10
Pygments==1.1.1
Werkzeug==0.9.6
bson==0.3.3
netaddr==0.7.13
pymongo==2.7.2
urlgrabber==3.9.1
urllib3==1.5
celery==3.1.17
python-crontab==1.9.2
gunicorn==19.2.1
python-digitalocean==1.4.1
linode==0.4
psutil==2.2.1
requests>=2.5.1
